fix: replace usize with i64 for COUNT(*) queries in rusqlite 0.38
rusqlite 0.38 removed FromSql/ToSql impl for usize by default.
Changed 4 test functions to use i64 instead.

+++ src-tauri/src/modules/db.rs
@@ -223,7 +223,7 @@ mod tests {
         let db_path = temp_dir.path().join("test.db");
         let db = Database::new_with_path(&db_path).unwrap();
 
-        let count: usize = db
+        let count: i64 = db
             .execute(|conn| {
                 Ok(conn.query_row("SELECT COUNT(*) FROM projects", [], |row| row.get(0))?)
             })
@@ -268,7 +268,7 @@ mod tests {
         })
         .unwrap();
 
-        let count: usize = db
+        let count: i64 = db
             .execute(|conn| {
                 Ok(conn.query_row("SELECT COUNT(*) FROM projects", [], |row| row.get(0))?)
             })
@@ -303,7 +303,7 @@ mod tests {
         assert!(result.is_ok());
 
         // Verify both records were committed
-        let count: usize = db
+        let count: i64 = db
             .execute(|conn| {
                 Ok(conn.query_row("SELECT COUNT(*) FROM projects", [], |row| row.get(0))?)
             })
@@ -340,7 +340,7 @@ mod tests {
         assert!(result.is_err());
 
         // Verify no records were committed
-        let count: usize = db
+        let count: i64 = db
             .execute(|conn| {
                 Ok(conn.query_row("SELECT COUNT(*) FROM projects", [], |row| row.get(0))?)
             })
